Allow the user to disable importing theano.
Note that importing theano can have side effects, like initializing/binding to the gpu.
Also the user should always be in control weather or not a optional dependency is tried to be loaded.
And if the user relais on configuring floatX through theano, failing to import theano should raise a exception instead of silently using adifferent default